Grocery Tracking Using FIREBASE DATATBASE
In India, the household food waste estimate is 50 kg per capita per year, or 68,760,163 tonnes a year. One of the reasons for this food waste is that the consumer forgets the expiry date of the food after purchase and does not consume the food before it expires, and when he/she finds out that the food is expired, it is tossed out. In the proposed system the user has to input some information regarding food products like product name, date of purchase, and date of expiry. On the basis of that data, an automatic notification will be sent to the customer’s smartphone several days before the purchased food expires. Also, it provides user to get recipes for breakfast, lunch and dinner. User can also search their nearby to order the expired food or any grocery they want.